Added GlobalExceptionHandler and 2 custom exception

+++ university-management/src/main/java/ua/develop/universitymanagement/service/impl/StudentServiceImpl.java
@@ -1,16 +1,18 @@
 package ua.develop.universitymanagement.service.impl;
 
 import lombok.AllArgsConstructor;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
 import ua.develop.universitymanagement.dto.student.StudentRead;
 import ua.develop.universitymanagement.dto.student.StudentSave;
+import ua.develop.universitymanagement.exceptions.custom.StudentAlreadyExistsException;
+import ua.develop.universitymanagement.exceptions.custom.StudentNotFoundException;
 import ua.develop.universitymanagement.mapper.StudentMapper;
 import ua.develop.universitymanagement.model.Student;
 import ua.develop.universitymanagement.repository.StudentRepo;
 import ua.develop.universitymanagement.service.StudentService;
 
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -22,6 +24,13 @@ public class StudentServiceImpl implements StudentService {
     @Override
     public Student saveStudent(StudentSave studentSave) {
         Student student = studentMapper.toStudent(studentSave);
+        if (studentRepo.getStudentByFirstNameAndLastNameAndPhoneNumber(
+                        studentSave.firstName(),
+                        studentSave.lastName(),
+                        studentSave.phoneNumber())
+                .equals(student)) {
+            throw new StudentAlreadyExistsException("");
+        }
         return studentRepo.save(student);
     }
 
@@ -40,13 +49,14 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void deleteStudentById(UUID id) {
-        try {
-            Student student = studentRepo
-                    .findById(id)
-                    .orElseThrow(ChangeSetPersister.NotFoundException::new);
-        } catch (ChangeSetPersister.NotFoundException e) {
-            throw new RuntimeException(e);
+        if (!studentRepo.existsById(id)) {
+            throw new StudentNotFoundException(id.toString());
         }
         studentRepo.deleteById(id);
+    }
+
+    private Student getStudent(UUID id) {
+        return studentRepo.findById(id)
+                .orElseThrow(NoSuchElementException::new);
     }
 }
